backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
django-cors-headers==3.14.0
djangorestframework-simplejwt==5.2.2
channels==4.0.0
Pillow==10.0.0
daphne==4.0.0